requirements.txt
aiohttp==3.5.4
aioredis==1.3.0
appdirs==1.4.3
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
bandit==1.5.1
black==19.3b0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.3.9
discord.py==1.2.5
entrypoints==0.3
flake8==3.7.7
flake8-polyfill==1.0.2
gitdb2==2.0.5
GitPython==2.1.11
greenlet==0.4.15
hiredis==1.0.1
idna==2.8
idna-ssl==1.1.0
mando==0.6.4
mccabe==0.6.1
more-itertools==7.0.0
msgpack==0.6.2
multidict==4.5.2
mypy==0.740
mypy-extensions==0.4.3
neovim==0.3.1
pbr==5.1.3
pluggy==0.9.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pynvim==0.3.2
pytest==4.4.1
pytest-asyncio==0.10.0
PyYAML==4.2b4
radon==2.4.0
requests==2.22.0
six==1.12.0
smmap2==2.0.5
stevedore==1.30.1
toml==0.10.0
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.25.6
websockets==6.0
xenon==0.5.5
yarl==1.3.0
